# Fix redundant first_set generation in closure