--- a/lr-core/src/lib.rs
+++ b/lr-core/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod grammar;
 pub mod lr;
+pub(crate) mod ordered_set;
 pub mod prelude;
 
 /// Represents the kind of table that can be generated