# Introduction
This PR moves first_set generation to the grammar table, where it was previously generated redundantly on each closure call. This value will never change once the grammar has been finalized, and is thus better suited to be done at the end of the grammar table initialization steps.

Initial tests from the benchmark show this alone leads to ~7x improvements to table set generation in a test grammar, and this number would expectantly grow as the grammar set increases.

## Test Results
before
```
Benchmarking lr1 table generation: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 75.0s, or reduce sample count to 10.
lr1 table generation    time:   [748.50 ms 753.76 ms 762.56 ms]
                        change: [+592.72% +597.77% +605.85%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) high mild
  7 (7.00%) high severe
```
after
```
Benchmarking lr1 table generation: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 10.6s, or reduce sample count to 40.
lr1 table generation    time:   [105.49 ms 105.55 ms 105.63 ms]
                        change: [-2.4185% -2.3277% -2.2396%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) low severe
  1 (1.00%) low mild
  2 (2.00%) high mild
  7 (7.00%) high severe
```

# Linked Issues
#41 
# Dependencies

# Test
- [x] Tested Locally
- [x] Documented

# Review
- [x] Ready for review
- [x] Ready to merge

# Deployment